requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
asyncer==0.0.7
attrs==25.3.0
auth0-python==4.9.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chainlit==2.5.5
chardet==3.0.4
charset-normalizer==3.4.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.120.1
crewai-tools==0.45.0
cryptography==44.0.3
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.9
embedchain==0.1.128
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.9
fastavro==1.10.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
gptcache==0.1.44
greenlet==3.2.2
grpcio==1.71.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
httpx-sse==0.4.0
huggingface-hub==0.31.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
instructor==1.8.2
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
json5==0.12.0
json_repair==0.44.1
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
lancedb==0.22.0
langchain==0.3.25
langchain-cohere==0.3.5
langchain-community==0.3.24
langchain-core==0.3.60
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langsmith==0.3.42
Lazify==0.4.0
litellm==1.68.0
literalai==0.1.201
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.9.0
mdurl==0.1.2
mem0ai==0.1.99
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.39.1
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.75.0
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.5
opentelemetry-instrumentation-anthropic==0.40.5
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-instrumentation-bedrock==0.40.5
opentelemetry-instrumentation-chromadb==0.40.5
opentelemetry-instrumentation-cohere==0.40.5
opentelemetry-instrumentation-crewai==0.40.5
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-instrumentation-google-generativeai==0.40.5
opentelemetry-instrumentation-groq==0.40.5
opentelemetry-instrumentation-haystack==0.40.5
opentelemetry-instrumentation-lancedb==0.40.5
opentelemetry-instrumentation-langchain==0.40.5
opentelemetry-instrumentation-llamaindex==0.40.5
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.5
opentelemetry-instrumentation-mcp==0.40.5
opentelemetry-instrumentation-milvus==0.40.5
opentelemetry-instrumentation-mistralai==0.40.5
opentelemetry-instrumentation-ollama==0.40.5
opentelemetry-instrumentation-openai==0.40.5
opentelemetry-instrumentation-pinecone==0.40.5
opentelemetry-instrumentation-qdrant==0.40.5
opentelemetry-instrumentation-replicate==0.40.5
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.5
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.5
opentelemetry-instrumentation-transformers==0.40.5
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.5
opentelemetry-instrumentation-watsonx==0.40.5
opentelemetry-instrumentation-weaviate==0.40.5
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.8
opentelemetry-util-http==0.52b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
portalocker==2.10.1
posthog==3.25.0
-e git+https://github.com/startlingadama/crewai-writer-system-agent-with-rag@34aa52506df2a21002cfe34de449fc8b5ee6990b#egg=projet
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.5.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.400
pysbd==0.3.4
pysqlite3-binary==0.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-multipart==0.0.18
python-socketio==5.13.0
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.14.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.25.0
rsa==4.9.1
schema==0.7.7
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.41.3
streamlit==1.45.1
sympy==1.14.0
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tornado==6.5
tqdm==4.67.1
traceloop-sdk==0.40.5
traitlets==5.14.3
typer==0.15.4
types-requests==2.32.0.20250515
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uptrace==1.31.0
urllib3==2.4.0
uv==0.7.4
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
